backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
langchain>=0.3.0
langchain-community>=0.0.10
langchain-openai>=0.3.3
langchain-chroma>=0.0.5
chromadb>=0.6.3
langchain-qdrant>=0.2.0
chroma-hnswlib>=0.7.3
BCrypt>=4.0.1
SQLAlchemy>=2.0.23
alembic>=1.12.1
mysql-connector-python>=8.0.33
minio>=7.2.0
python-docx>=0.8.11
pypdf>=3.0.0
unstructured>=0.10.30
markdown>=3.0.0
unstructured[md]>=0.10.0
openai>=1.30.0
email-validator
dashscope>=1.13.6
langchain-deepseek==0.1.1
langchain-ollama==0.2.3
docx2txt==0.8
numpy>=1.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pyyaml>=6.0.0
faiss-cpu>=1.7.4
pinecone-client>=3.0.0